Switch from using Node and npx to bun and bunx
- Bun can directly execute TypeScript code instead of needing ts-node
  - Overall execution also seems slightly faster
- Add `package.json` for dev-dependency on `@types/node`
- Move types directly into script and fix type errors
- Remove `npm-cache-dependency-path` parameter in action inputs
