+++ action.yml
@@ -9,9 +9,6 @@ inputs:
   workspace:
     required: true
     description: The workspace to filter on
-  npm-cache-dependency-path:
-    description: Value given for cache-dependency-path to actions/setup-node
-    default: ""
 outputs:
   status:
     description: "Status is 0 if changes can be ignored, 1 if changes should be deployed."
@@ -30,12 +27,10 @@ runs:
         # This comparison also assumes the previous deployment succeeded
         fetch-depth: 2
 
-    - name: Set up Node
-      uses: actions/setup-node@v4
+    - name: Set up Bun
+      uses: oven-sh/setup-bun@v2
       with:
-        node-version: 20
-        cache: "npm"
-        cache-dependency-path: ${{ inputs.npm-cache-dependency-path }}
+        bun-version: latest
 
     - name: Run script
       id: run_script
@@ -45,5 +40,5 @@ runs:
         WORKSPACE: ${{ inputs.workspace }}
       shell: bash
       run: |
-        v=$(node ${{ github.action_path }}/super-turbo-ignore.js $TASK $WORKSPACE)
+        v=$(bun ${{ github.action_path }}/super-turbo-ignore.ts $TASK $WORKSPACE)
         echo "status=${v}" >> $GITHUB_OUTPUT
