+++ super-turbo-ignore.ts
@@ -1,10 +1,9 @@
 // Derived from vercel/turborepo:packages/turbo-ignore/src/ignore.ts@29aeaf0
 // See MIT License from Vercel, Inc
+// Keeping in a single file to keep this portable and help with testing externally
 
-/** @import { CommitCheck, PackageJson, DryRun } from './types' */
-
-const { execSync } = require("node:child_process");
-const fs = require("node:fs");
+import { execSync } from "node:child_process";
+import fs from "node:fs";
 
 // the "≫" symbol
 const LOG_PREFIX = "\u226B  ";
@@ -21,12 +20,16 @@ function output(...args) {
 	console.log(...args);
 }
 
+interface CommitCheck {
+	result: "deploy" | "continue";
+	reason: string;
+}
+
 /**
+ * Check commit messages for indicator to force deployment
  * Adapted from turbo-ignore/src/checkCommit.ts@257d0c0
- * @param {string} workspace - the workspace involved
- * @returns {CommitCheck}
  */
-function checkCommit(workspace) {
+function checkCommit(workspace: string): CommitCheck {
 	const commitMessage = execSync("git show -s --format=%B").toString();
 
 	const forceWorkspaceDeploy = `[vercel deploy ${workspace}]`;
@@ -43,15 +46,21 @@ function checkCommit(workspace) {
 	};
 }
 
+type Dependencies = Record<string, string>;
+
+interface PackageJson {
+	dependencies: Dependencies;
+	devDependencies: Dependencies;
+}
+
 /**
+ * Read turbo version from package.json
  * Adapted from turbo-ignore/src/getTurboVersion.ts@29aeaf0
- * @returns {string}
  */
-function getTurboVersion() {
+function getTurboVersion(): string {
 	try {
 		const raw = fs.readFileSync("package.json", "utf8");
-		/** @type {PackageJson} */
-		const packageJson = JSON.parse(raw);
+		const packageJson = JSON.parse(raw) as PackageJson;
 		const dependencies = packageJson.dependencies?.turbo;
 		const devDependencies = packageJson.devDependencies?.turbo;
 		const turboVersion = dependencies || devDependencies;
@@ -59,21 +68,32 @@ function getTurboVersion() {
 			info(`Inferred turbo version "${turboVersion}" from "package.json"`);
 			return turboVersion;
 		}
+		return "latest";
 	} catch (e) {
-		error(`"${packageJsonPath}" could not be read`);
+		error(`"package.json" could not be read`);
 		throw e;
 	}
+}
+
+interface Task {
+	taskId: string;
+	task: string;
+	package: string;
+	hash: string;
+}
+
+interface DryRun {
+	packages: string[];
+	tasks: Task[];
 }
 
 /**
  * Check if the deployment can be skipped by comparing with the previous commit
  * Allows forcing deployment from commit message
  * Assumes package has lockfile, package manager is set, and commit has reachable parent
- * @param {string} workspace - the workspace to use
- * @param {string} task - the task to run
  * @returns {boolean} whether or not the deployment can be skipped
  */
-function superTurboIgnore(task, workspace) {
+function superTurboIgnore(task: string, workspace: string): boolean {
 	info(`Using Turborepo to determine if this project is affected by the commit...\n`);
 
 	// check for TURBO_FORCE and bail early if it's set
@@ -83,7 +103,7 @@ function superTurboIgnore(task, workspace) {
 	}
 
 	// check the commit message for force deploy
-	const parsedCommit = checkCommit({ workspace });
+	const parsedCommit = checkCommit(workspace);
 	if (parsedCommit.result === "deploy") {
 		info(parsedCommit.reason);
 		return false;
@@ -91,7 +111,7 @@ function superTurboIgnore(task, workspace) {
 
 	const turboVersion = getTurboVersion();
 	const turbo = `turbo@${turboVersion}`;
-	const command = `npx -y ${turbo} run ${task} --filter="${workspace}" --dry=json`;
+	const command = `bunx ${turbo} run ${task} --filter="${workspace}" --dry=json`;
 	// Assume cwd is the monorepo root
 	const cwd = process.cwd();
 
@@ -99,10 +119,9 @@ function superTurboIgnore(task, workspace) {
 		info(`Analyzing results of \`${command}\``);
 
 		try {
-			const stdout = execSync(command, { cwd });
+			const stdout = execSync(command, { cwd, encoding: "utf8" });
 
-			/** @type {DryRun} */
-			const parsed = JSON.parse(stdout);
+			const parsed = JSON.parse(stdout) as DryRun;
 			const { packages, tasks } = parsed;
 
 			if (!packages.includes(workspace)) {
@@ -132,7 +151,7 @@ function superTurboIgnore(task, workspace) {
 			return false;
 		}
 
-		const hashChanges = [];
+		const hashChanges: string[] = [];
 		// Order should be deterministic by top-sort, so check pairs
 		for (let i = 0; i < currentTasks.length; ++i) {
 			const currentTask = currentTasks[i];
